requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching & Message Queue
redis==5.0.1
hiredis==2.3.2

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Monitoring & Metrics
prometheus-client==0.19.0
evidently==0.4.14

# Data Quality
great-expectations==0.18.8

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Workflow Orchestration
apache-airflow-client==2.8.0

# Kubernetes & Helm
kubernetes==29.0.0
python-helm==0.1.0

# Graph Processing
networkx==3.2.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.3
factory-boy==3.3.0
faker==22.2.0
testcontainers==3.7.1
locust==2.20.0

# Development Tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0
types-redis==4.6.0.20240106
types-pyyaml==6.0.12.12
